Users can view private post
