+++ src/components/ApplicationViews.js
@@ -1,11 +1,9 @@
 import React from "react"
 import { Route } from "react-router-dom"
 import { ShowCategories } from "./categories/Categories"
-import { PostList } from "./posts/Posts"
-import { PostDetails } from "./posts/PostDetails"
+import { PostList } from "./posts/SharedPost"
 import { PostForm } from "./posts/CreateNewPost"
 import { UsersList } from "./users/Users"
-import { MyPosts } from "./posts/MyPosts"
 import { UserDetails } from "./users/UserDetails"
 import { ShowTags } from "./tags/Tag"
 import { UpdateTag } from "./tags/UpdateTag"
@@ -22,14 +20,9 @@ export const ApplicationViews = () => {
 				{/* child */}
 				<PostList />
 			</Route>
-			<Route exact path='/MyPosts'>
-				<MyPosts />
-			</Route>
 			{/* when the url is posts, display postId- capturing after : and storing */}
 			{/* postId is the key post component! */}
-			<Route exact path='/posts/:postId(\d+)'>
-				<PostDetails />
-			</Route>
+
 			<Route exact path='/posts/:postId(\d+)/update'>
 				<UpdatePost />
 			</Route>
